fix build warnings with type mismatches and base check helpers
Building with base headers in a SLOT-ed setup exposes build warnings
(which causes failures due to -Werror).  One such example:

In file included from extent_writer.h:9:0,
                 from bzip_extent_writer.h:10,
                 from bzip_extent_writer.cc:5:
.../base/logging.h: In function 'std::string* logging::CheckEQImpl(const t1&, const t2&, const char*)
                                 [with t1 = unsigned int, t2 = int, std::string = std::basic_string<char>]':
bzip_extent_writer.cc:53:7:   instantiated from here
.../base/logging.h:512:1: error: comparison between signed and unsigned integer expressions [-Werror=sign-compare]
cc1plus: all warnings being treated as errors

Explicitly cast the constants to avoid these.

BUG=chromium-os:16623
TEST=`emerge-x86-alex update_engine` builds

Reviewed-on: https://gerrit.chromium.org/gerrit/16035
Tested-by: Mike Frysinger <vapier@chromium.org>
Commit-Ready: Mike Frysinger <vapier@chromium.org>

+++ bzip_extent_writer.cc
@@ -50,7 +50,7 @@ bool BzipExtentWriter::Write(const void* bytes, size_t count) {
                      output_buffer.size() - stream_.avail_out));
     
     if (rc == BZ_STREAM_END)
-      CHECK_EQ(stream_.avail_in, 0);
+      CHECK_EQ(stream_.avail_in, static_cast<unsigned int>(0));
     if (stream_.avail_in == 0)
       break;  // no more input to process
   }
